chore: dispatch in app review on successful swap and reward claim (#3768)
### Description

Adds the additional trigger actions of `swapSuccess` and
`claimRewardsSuccess` for In App Review. I opted to excluded dapp tx as
the action in-app review would likely disrupt the dapp action flow and
this is against [best
practices](https://developer.apple.com/design/human-interface-guidelines/ratings-and-reviews).

> Avoid interrupting people while they’re performing a task or playing a
game. Asking for feedback can disrupt the user experience and feel like
a burden. Look for natural breaks or stopping points in your app or game
where a rating request is less likely to be bothersome.



### Test plan

Tested locally on iOS.

### Related issues

- Fixes RET-15

### Backwards compatibility

Yes