fix(home-actions): use X instead of back for some actions (#3813)
### Description

The animation for back chevron is broken on the QR screen. After talking
to Nitya, we decided to revert the icon to X on the send, receive (QR)
and request actions. Swap, Add and Withdraw still use back. Also reverts
the Animated change on the BackChevron component since it wasn't working
anyway.

### Test plan

Unit tests, manual


https://github.com/valora-inc/wallet/assets/5062591/cf2e5816-f05d-4246-9047-489f6f6788f2


### Related issues

- Fixes ACT-811

### Backwards compatibility

Yes