fix: android release build failure (#3910)
### Description

This PR fixes the build issue we see
[here](https://github.com/valora-inc/release-automation/actions/runs/5350577625/jobs/9708127671#step:17:95).
it seems to indicate that there were some changes in the maven
dependencies.

This PR:
- applies
https://github.com/EdgeApp/react-native-fast-crypto/commit/35f2106bb6904afc4e936b201ff1c46afd56b3f5
as a patch. unfortunately i tried* but failed to make the latest version
of react-native-fast-crypto compatible with android 6 and 7, so in the
interest of time the patch seems the best option.
- upgrades the exoplayer versions - after resolving the fast-crypto
error, i saw the build error below. the version added here is what is
recommended in the clevertap-react-native
[docs](https://developer.clevertap.com/docs/react-native-quick-start-guide#add-dependencies)
(we didn't update this part when renovate upgraded the project)
```
FAILURE: Build failed with an exception.

* What went wrong:
Could not determine the dependencies of task ':app:compileAlfajoresdevDebugJavaWithJavac'.
> Could not resolve all task dependencies for configuration ':app:alfajoresdevDebugCompileClasspath'.
   > Could not find com.google.android.exoplayer:exoplayer:2.11.5.
     Required by:
         project :app
   > Could not find com.google.android.exoplayer:exoplayer-hls:2.11.5.
     Required by:
         project :app
   > Could not find com.google.android.exoplayer:exoplayer-ui:2.11.5.
     Required by:
         project :app
```

*The things i tried:
- for android API level below 26, use PBKDF2WithHmacSHA1 instead of
PBKDF2WithHmacSHA512 - this has security implications so it would have
been quick and dirty, but it didn't work anyway (downstream functions
did not accept the output)
- for android API level below 26, put back all the code that was remove
in [this
commit](https://github.com/EdgeApp/react-native-fast-crypto/commit/91b4af45c872f2f23029a600e3d1b243ee2571bf)
- maybe i didn't do this well, but it just caused an app crash for me.
unfortunately this meant i have no idea what went wrong, crash happens
too fast to see anything in live logs and can't send logs to myself
because no email client on the emulator.

### Test plan

Did a build locally, verified working on android 7 emulator

### Related issues

n/a

### Backwards compatibility

Y