feat: remove the legacy send payment saga (#3921)
### Description

Currently, there is both a legacy and regular send payment saga in the
wallet. This PR removes the legacy saga and any other code that was a
part of this legacy flow. Moreover, this PR updates handling payment
request notifications to navigate to the regular send payment flow
instead of the legacy one.

##### Legacy Send Confirmation Screen To Be Removed


![SendConfirmationLegacyExample](https://github.com/valora-inc/wallet/assets/20508929/89f0beb3-c1b1-47c7-9905-b3899757ea66)


##### Existing Send Confirmation Screen


![SendConfirmationExample](https://github.com/valora-inc/wallet/assets/20508929/d1e343b9-8b6d-4de8-8f91-f2ae58865b58)


### Test plan

Added and updated unit tests.

### Related issues

- ACT-750

### Backwards compatibility

Yes

---------

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>
Co-authored-by: Jacob Waterman <jacobrwaterman@gmail.com>