feat(wallet): Genericize ValoraWallet interface (#3967)
### Description

Currently, the `ValoraWallet` interface expects Celo-specific types to
be used for the `signTransaction` method. The input/output types are not
easily transformed between Ethers/Celo-specific types, however, which
means that we will not be able to implement `ValoraEthersWallet` using
the Celo types (or modify `ValoraCeloWallet` to use Ethers types). To
sidestep this, this PR genericizes the `ValoraWallet` interface such
that either Celo _or_ Ethers types can be used for `signTransaction`.
When we decide to switch the primary wallet within the app to a
`ValoraEthersWallet` implementation, all we'll need to do is change a
single type (`PrimaryValoraWallet`) and update the call sites of
`signTransaction` throughout the app to use Ethers-specific types. Since
all the other method signatures are shared between `ValoraCeloWallet`
and `ValoraEthersWallet`, these are the only changes we'll need to make.

### Test plan

Manual tested

### Related issues

### Backwards compatibility

Yes.