feat(wc): switch to web3wallet (#3955)
### Description

Switches the wallet connect library used from
`@walletconnect/sign-client` to `@walletconnect/web3wallet` which
supports use cases that will be available for builders going forward,
that the sign-client currently does not support.

`getSessionFromRequest` and `handleAcceptRequest` have the largest
changes where instead of iterating over an array of sessions from redux
to find the session they use `client.getActiveSessions()` which produces
an object from which the session in retrieved. In testing locally
occasionally the existing setup would not produce an action request when
expected; however, using the client singleton to get the matching
session has worked consistently.

### Test plan

- [x] Existing e2e tests updated to use an updated sign-client for the
sample dapp and passing
- [x] Existing unit tests passing.
- [x] Tested locally on iOS and Android with:
  - https://react-app.walletconnect.com
  - https://www.impactmarket.com
  - https://app.stcelo.xyz
  - https://mento.finance
  - https://churrito.fi

### Related issues

- Fixes RET-756

### Backwards compatibility

Yes, but users will likely need to reconnect to dapps.