feat: add confirmation request when claiming shortcut rewards (#3977)
### Description

This PR adds confirmation bottom sheet to reward shortcuts screen. The
confirmation bottom sheet displays the raw data for all the transactions
that will be sent. On confirm, all the transactions will be sent.
1. previously, we were fetching and executing the shortcut transactions
with one action. now, we split them up. `triggerShortcut` fetches the
shortcut transaction, we wait for the user's approval, then
`executeShortcut` submits the transaction to the blockchain.
2. on click of the "claim" button, mark the reward as "loading" -
display the loader on the claim button of the reward with this. start to
fetch the shortcut transactions.
3. on receiving the shortcut transactions, store them in redux and mark
the reward as "pendingAccept"
4. if the user allows the transaction, mark the reward as "accepting" -
display the loader on the bottom sheet with this and begin to send the
transactions.
5. if the user denies the transaction by dismissing the bottom sheet,
mark the reward as "idle".


### Test plan


Show bottom sheet on claim -> copy tx data -> show what happens when the
send transaction fails. Show what happens when the user has already
approved the transaction and dismisses the bottom sheet (the tx should
continue, it is hard coded to fail for this example)



https://github.com/valora-inc/wallet/assets/20150449/507de39a-cad6-4162-bdb0-6c878664c017



On success, the bottom sheet should dismiss itself. The reward should be
marked as claimed, and refreshed with claim enabled if the amount
changes.



https://github.com/valora-inc/wallet/assets/20150449/146c8cd3-b07a-4811-b17b-181af9a52b7e




### Related issues

- Fixes RET-733

### Backwards compatibility

Y