fix(wc): better compat with unsupported chains (#3998)
### Description

This improves connecting to dapps which either don't specify the Celo
chain in the required namespaces or have it only in the optional
namespaces.

Curve, Toucan and Cred Protocol fall into this.
The typical issue was that the dapp was reporting the wallet couldn't
switch network.
Specifically with Curve, the Ethereum network was selected in the
dropdown and it was impossible to switch it.

Previously we were allowing any chain. Now we only only EVM chains, even
if not Celo.
We then clearly warn if the dapp tries to perform an action on an
unsupported chain, and reject it.

I've left a note to also update the connect experience to show a warning
when the requested chain/method/event aren't supported. This will be
done in a follow up PR.

### Test plan

- Updated tests
- Manually tested connecting to all listed dapps.

Example of the new warning when the chain is not supported (note: this
isn't a real world example, I manually changed the values to produce
this):

<img
src="https://github.com/valora-inc/wallet/assets/57791/738bdad2-ef31-4d32-a11b-716b67267977"
width="50%" />

### Related issues

- Fixes RET-778

### Backwards compatibility

Yes.