feat(cab): Add KeylessBackupProgress screen (#4018)
### Description

For ACT-763. Adds a progress screen for the actual backup/restore
portion, after both auth methods are successful. See Figma designs
[here](https://www.figma.com/file/gGp6SVFWMbnSltkGzNzrpC/Wallet-Recovery?type=design&node-id=3604-31351&mode=dev).

This adds in a new piece of Redux state to track the progress of the
backup, and dispatches a (currently dummy) new action to kick off a saga
to begin the backup process. This was all designed with the restore flow
in mind, so should be able to be re-used for restore, if we wish.

### Test plan

Unit and manual tested. See below images:


![2023-07-31-215926_829x1733_scrot](https://github.com/valora-inc/wallet/assets/569401/7f87e82d-47aa-4b44-89f7-99de9426607f)

![2023-07-31-215907_824x1737_scrot](https://github.com/valora-inc/wallet/assets/569401/274a17d3-54de-4555-aa03-edda563bb8ec)

![2023-07-31-215842_823x1734_scrot](https://github.com/valora-inc/wallet/assets/569401/22376ac2-1228-454a-aaf1-9840c119a7fe)

### Related issues

- Fixes ACT-763.

### Backwards compatibility

N/A.