feat(notifications): flatten payment request notifications (#4075)
### Description

Currently, we have one "summary" notification card that groups multiple
incoming payment requests. On tap of this card, a user can enter a
screen that shows a list with each individual request.

It is a similar flow for outgoing payment requests and sent escrow
invites.

In this PR each payment request is displayed directly on the
notification feed, so there is no extra screen transition required to
interact with individual notifications.

Within their priority group, payment requests are sorted in reverse
chronological order (the most recent on top).


![flatten-notifications](https://github.com/valora-inc/wallet/assets/2737872/432aec60-466c-40cb-941f-d1e1f94222dc)

#### Minor fix

For the sake of consistency, the incoming payment request card is
updated to display either a phone number or the address of the requester
if no name is associated with the address. It is the same behavior as in
outgoing payment requests.

### Test plan

* Tested manually on mocked data on iOS and Android emulators.
* Unit tests updated to ensure correct ordering of requests and correct
render of incoming payment card

### Related issues

- Fixes RET-803

### Backwards compatibility

N/A