refactor(analytics): avoid duplication in defining AnalyticsEventType (#4119)
### Description

previously there were a couple of event type enums that were missing
from AnalyticsEventType, including `QrScreenEvents` and
`PhoneVerificationEvents`. This was possible because
`ValoraAnalytics.track` uses `keyof AnalyticsPropertiesList` as an input
type, not `AnalyticsEventType`. There are a few places we do actually
use `AnalyticsEventType`, so I didn't get rid of it; instead I changed
it to `keyof AnalyticsPropertiesList`, taking a cue from the typing on
`ValoraAnalytics.type`, be easier to keep in sync with the other
analytics types.

### Test plan

typescript compiles

### Related issues

na

### Backwards compatibility

na