fix: ensure redux selectors are properly memoized (#4124)
### Description

During the RN upgrade I noticed some errors in the test logs `Selector
walletConnectEnabledSelector returned a different result when called
with the same parameters. This can lead to new ref. Selectors that
return a new reference (such as an object or an array) should be
memoized:
https://redux.js.org/usage/deriving-data-selectors#optimizing-selectors-with-memoization`.
This PR fixes the selectors that were complained about.


### Test plan

n/a

### Related issues

n/a

### Backwards compatibility

Y

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>