docs(analytics): compile-time enforcement of analytics docs (#4120)
### Description

relies on #4119 

Thinking of creating a new analytics event without documenting it and
leaving your poor teammates in the dark about what it does? Think again!
😈

The main contribution is `eventDocs`, which is typed to require an entry
for analytics event we define. This gives us a CI check for when we add
analytics events but fail to document them (since typescript will not
compile).

For events with no docs, I used an empty string as the description 😢.
For events with inline code comments, I used the code comment as the
description (using a script).

I then deleted the inline code comments in Events.tsx to avoid having 2
sources of truth and confusing teammates who rely on these docs.

Document your analytics events!!!

### Test plan

typescript compiles

spot checked many of the event descriptions

### Related issues

na

### Backwards compatibility

na