feat(nav-drawer): upgrade to v6 and enable freeze (#4135)
### Description

Updates @react-navigation/drawer to v6 and enables freeze for
performance improvements. Some prop changes were needed to maintain
existing styles and behavior.

Freeze works as follows:
> ...Using Suspense mechanism introduced in React 17. The main use case
of this library is to avoid unnecessary re-renders of parts of the app
that are not visible to the user at a given moment. The frozen
components are not unmounted when they are replaced with a placeholder
view, so their React state and corresponding native view instances are
retained during the freeze (DOM elements for react-dom or
platform-native views for React Native apps) keeping things like scroll
position, input state, or loaded images (for <img> components)
unchanged. [source](https://github.com/software-mansion/react-freeze).

| iOS Before | iOS After | Android Before | Android After |
| :---: | :---: | :---: | :---: |
|
![](https://github.com/valora-inc/wallet/assets/26950305/3375b1e1-193c-4f84-b653-b17db439262f
"iOS Before") |
![](https://github.com/valora-inc/wallet/assets/26950305/804d2ddf-7fff-461a-82f0-288944ca5b63
"iOS After")|
![](https://github.com/valora-inc/wallet/assets/26950305/81d12fc3-fa9b-4467-9143-b429946b538d
"Android Before") |
![](https://github.com/valora-inc/wallet/assets/26950305/5bcd5206-3713-47a7-841d-3b7407978e71
"Android After") |

### Test plan

- Tested locally on Android & iOS

### Related issues

N/A

### Backwards compatibility

Yes