feat(multichain): Render multichain transfers on the home feed (#4115)
### Description

For ACT-887. This PR updates the wallet to render non-Celo transfers in
the transaction feed on the home screen. This feature is locked behind a
dynamic config parameter at the moment ([this
one](https://console.statsig.com/4plizaPmWwPL21ASV4QAO0/dynamic_configs/multi_chain_features),
to be exact). This dynamic config also contains a parameter for
separately controlling rendering of native tokens. For users who fail
the feature gate, network requests will _not_ be made to fetch non-Celo
transfers, and any previously cached non-Celo transfers will not be
rendered.

The actual rendering of non-Celo transfers is kind of broken at the
moment, but that's to be expected. In the video below, you can see the
example transfer on Goerli Etherscan
[here](https://goerli.etherscan.io/address/0xe192673ad298ff0b72dd1c8e074d34ddaafd9591)
(compare the send address against the video).

Note that actually dealing with native assets (i.e., token-like thing
that lacks an actual address) is being punted on for now, since it will
require a larger overhaul to how we represent tokens in the wallet and
throughout our entire stack.

### Test plan

Manual tested on the Goerli testnet. See video below.


https://github.com/valora-inc/wallet/assets/569401/878a193f-e1f4-4b62-9fc2-1360625a4537

### Related issues

- Fixes ACT-887.

### Backwards compatibility

Yes.