feat(networks): Use `networkId` at the API level (#4192)
### Description

Currently in draft state; the actual API changes required for this to
work have not been merged yet. This depends on changes to the APIs in
`cloud-functions` and `blockchain-api`.

* `cloud-functions` PR implementing the required changes:
[here](https://github.com/valora-inc/cloud-functions/pull/129)
* `blockchain-api` PR implementing the required changes: TBD

For some context, we want to get this change in before the release cut.
If we ever want our backend services to be able to speak to multiple
different blockchain environments, _without this change_, we'll need a
bit of a hacky workaround. (e.g., if we want production GCP
blockchain-api to be able to fetch data from `celo-alfajores` or
`ethereum-sepolia`.) If we do not make this change, and want to have
this kind of support in the future, we'll need to update
`blockchain-api` and `fetchProvider` in a backwards compatible way in
order to not break intermediate app versions. This sort of BC change
would be somewhat hacky - there are two obvious ways we could do it:

* Have the existing `network` parameter in these APIs expanded or
"overloaded" to recognize _both_ env-agnostic and env-specific network
identifiers (with env-agnostic identifiers "falling back" to the
service's deploy environment to resolve the correct env-specific
identifier).
* Adding another, optional, env-specific parameter to these endpoints,
so they can field requests both for env-agnostic identifiers and
env-specific identifiers. This would mean it would be possible to call
these endpoints with:
* no network identifier at all (falling back onto celo as default chain
+ deploy env)
  * an env-specific identifier
  * an env-agnostic identifier (falling back onto deploy env)
* both an env-specific and env-agnostic identifier (undefined behavior?)

Since we've only just really begun hooking together these multi-chain
backend APIs with the wallet, it doesn't make a ton of sense to
introduce the potential for a hacky-ish fix down the road when we can
obviate the need for it now.

### Test plan

Unit tests updated; will manual test on Alfajores once the
`cloud-functions` and `blockchain-api` PRs have been merged and
deployed.

### Related issues

N/A

### Backwards compatibility

Yes.

---------

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>