feat: make TokensBottomSheet use native bottom sheet (#4244)
### Description

This PR updates TokenBottomSheet to use the (native) BottomSheet
component. TokenBottomSheet was used by SendAmountHeader, Send, and
SwapScreen. Calling out changes related to each path:
1. Send: this was the most straightforward. The component now uses a ref
instead of state variable to use the bottom sheet.
- i removed a snapshot test because it didn't seem to be adding value
and would have bloated this PR with hundreds of new lines of changes.
2. SendAmountHeader: the token selection button exists on the header,
but the bottom sheet should not be triggered from the header (it causes
hierarchy issues where the bottom sheet is rendered underneath the
actual screen content). It made more sense to shift the bottom sheet to
the SendAmount/index.js file. The changes here helped to simplify both
components I think.
3. SwapScreen - this required the most changes.
- to get the bottom sheet background to display on top of the screen
header, i replaced the nav header with the CustomHeader inside the
component directly.
- there was some leftover code from the home screen quick actions
experiment which was causing complexity so i removed them.
- update BottomSheet to accept a sticky header component, for the search
bar
- update BottomSheet to accept snap points. the
[designs](https://www.figma.com/file/YxVf7jqvALnHfzr1Vgz5uV/Swap?type=design&node-id=628-19734&mode=design&t=EWzb6qN2FqgvzIEt-0)
show that the bottom sheet should not take up the full screen height.
- i spent hours trying to get the sticky header to work nicely with
dynamic heights. i could not do it, and left a comment in the component.
we don't have a use case for this yet, so i don't want to spend more
time on it.

### Test plan

iOS:


https://github.com/valora-inc/wallet/assets/20150449/4c482573-672f-4453-ac61-1e2250f1778b

Android:


https://github.com/valora-inc/wallet/assets/20150449/e4c1d5d1-1640-460d-a22a-88119610b2fc

### Related issues

- Fixes RET-871

### Backwards compatibility

Y