fix: back button hit area too small (#4304)
### Description

Increased back button hit area, which was too small and almost
impossible to hit.

I'm not too happy about the manual padding added, I would have preferred
the
[hitSlop](https://github.com/valora-inc/wallet/blob/ede52965f035ce263c38e83839be64c7412cb3bb/src/navigator/TopBarButton.tsx#L47)
to work, but somehow it doesn't.
We can investigate this more later. But for now this should be good
enough.

### Test plan

**before/after**

Showing with the ripple effect:
<img
src="https://github.com/valora-inc/wallet/assets/57791/f3dd90dc-c717-420e-b7e1-bccb1e78ee6b"
width="50%" /><img
src="https://github.com/valora-inc/wallet/assets/57791/56b7f956-61aa-410e-9982-1dc6407edca2"
width="50%" />

Note: it would be nice for the ripple effect to not be cut off on the
left, but this would mean moving the button more toward the right, and
then not be aligned anymore with the content. We can discuss that with
design later on.

### Related issues

- See Slack
[thread](https://valora-app.slack.com/archives/C025V1D6F3J/p1696953791346979?thread_ts=1696932462.287309&cid=C025V1D6F3J)

### Backwards compatibility

Yes