fix: show completed standby transactions (#4380)
### Description

There are a few things going on in this PR to achieve the end result:
1. update the TransactionFeed component to show transactions from
blockchain-api _and_ the completed standby transactions. to achieve
this, the standby transaction should have similar data schema.
2. update the `StandbyTransaction` type to make step 1 easier: when a tx
is confirmed, use the tx receipt to make the standby tx data as similar
as possible to what comes back from blockchain-api from the reducer. i
haven't created the fees object in this PR which means that between the
tx confirmed and blockchain-api catching up, no fees can be shown in the
tx details screen.
4. i removed the TRANSACTION_CONFIRMED_VIEM action because it seems to
do the same thing as TRANSACTION_CONFIRMED. rather than rewriting the
same reducer twice, i just combined the actions.

I'll handle the fees in a separate PR.


### Test plan

Before this change:


https://github.com/valora-inc/wallet/assets/20150449/f349eec4-6011-4335-a323-de132ea15e66

After this change:


https://github.com/valora-inc/wallet/assets/20150449/6c585c10-467f-4790-9391-c05e1c7160a7


### Related issues

- Fixes RET-894

### Backwards compatibility

Y