feat(sends): Support sending native/ERC20 non-Celo assets using Viem (#4306)
### Description

For ACT-904. Adds support for e2e sends on Ethereum, including both the
native ETH asset and arbitrary ERC20 tokens.

A couple of notes:
* Supporting non-Celo sends requires _two_ separate ways of submitting
TXs to the blockchain. For ERC20 sends, where we definitely have a
contract address, we simply re-use the existing contract
simulation/request logic, extending it to Viem clients for other chains.
For native sends, we don't have a contract address available, so we must
simulate the send using a `call`, followed by a `sendTransaction`.
* Since we don't have Ethereum fee estimation, fees have been made
optional for Ethereum for the time being, while still being required for
all sends on Celo.

### Test plan

Unit and manual tested. Note that in the videos below, the standby
transaction is added correctly, but since I'm sending the assets to my
own address, blockchain-api sees that transaction as containing exactly
1 send and 1 receive, so classifies it as a swap in the TX feed once the
TX has been confirmed.

#### Native ETH Send


https://github.com/valora-inc/wallet/assets/569401/306d0fca-1721-4f83-a613-5c38db8931b0

#### ERC20 USDC Send


https://github.com/valora-inc/wallet/assets/569401/7d2973ba-355e-4ff4-b8c9-54074dc8e586

### Related issues

- Fixes ACT-904

### Backwards compatibility

Yes.

---------

Co-authored-by: MuckT <mcgtom10@gmail.com>
Co-authored-by: Charlie Andrews <charlie.andrews@valoraapp.com>
Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>