fix: ensure transactions for all networks are correctly saved in redux (#4388)
### Description

Context
https://valora-app.slack.com/archives/C025V1D6F3J/p1698145522738719

As the title. Currently we are writing and overwriting the transactions
redux store constantly. The issue is happening because blockchain-api
returns transactions per network, and we save these transactions for a
certain network (replacing the existing transactions) to redux. If there
is more than one network enabled, the result is constant overwriting
since we are reading the transaction hashes in the redux store to decide
whether there are new transactions....

The fix:
- make redux schema reflect how the data that we receive, since it acts
like a cache layer
- use selectors to transform the data into what we want for display

### Test plan

Before:


https://github.com/valora-inc/wallet/assets/20150449/5ac66014-5dc1-45e9-9ab3-8ba9e12d6e54

After, app start without any cached tx's:


https://github.com/valora-inc/wallet/assets/20150449/10497e11-c806-41ec-bba6-7423d7a4d532

After, receiving 2 transactions on different networks. Note the
timestamps of the actions dispatched (only once, when the transaction is
returned for the first time).


https://github.com/valora-inc/wallet/assets/20150449/ff696f46-62ef-40f6-b7b9-5db4f99eff7a


### Related issues

n/a

### Backwards compatibility

Y