revert: cash in bottom sheet removal (#4446)
### Description

Reverts the changes from #4383

We want to keep it for now, to be able to switch it back on in the
future. See Slack
[thread](https://valora-app.slack.com/archives/C02898GN22V/p1699383234679679?thread_ts=1698272444.022329&cid=C02898GN22V).

### Test plan

- [x] Redux migrations updated
- [x] Existing e2e tests restored
- [x] Existing unit tests restored

### Related issues

N/A

### Backwards compatibility

Users should not have an issue upgrading from 1.70.4 to the next
release; however, developers will need a fresh install due to the
changes in the redux migrations.