feat: add network fee to swap screen (#4425)
### Description

This PR:
- add network fee in fiat currency to swap tx details display
- enable swap tx details display on the swap input screen permanently

Not included in this PR:
- skeleton loading state for the network fee

For the network fee value, I tried to emulate the logic used in the swap
review screen but since we are axing the contractkit flow soon, i didn't
want to copy over too much extra logic.
- for the viem swap flow, use the fees in the prepared transactions. if
the transaction isn't possible, default to "-" (the user can't proceed
with the swap anyway). for ease of use, i added the total fee to the
prepared transaction, since it is already there.
- for the contractkit flow, i made an approximation using the data
provided by the quote itself and assuming that the fee currency is Celo
(rather than bringing over all the fee currency calculation logic).
since gas fees are low, and we are displaying it in fiat value, i think
the difference will be negligable.
- use the TokenDisplay component to show the fiat value of the network
fee, rather than calculating it upfront.

### Test plan

Using the viem and contractKit flows, verify that the network fee value
displayed in this PR is the same as what is displayed in the swap review
screen. **Note that there are some performance issues on this screen - I
can see the same issues on main so will address those separately, they
were not introduced by changes in this PR.** At the time of recording,
our swap providers are having performance issues so the videos are very
long :(

ContractKit: notice the fee currency in fiat in the first screen is the
same as the fee currency in token units since i have EUR local currency
+ cEUR as the fee currency.


https://github.com/valora-inc/wallet/assets/20150449/a9638538-4205-434d-8de7-196609772b9e


Viem: notice the fee currency in fiat is half as much as the token value
in the swap review screen (1 CELO is roughly 0.5EUR)



https://github.com/valora-inc/wallet/assets/20150449/f41bcbd9-dc82-4839-964d-bf9c47d9a781


### Related issues

- Related to RET-865

### Backwards compatibility

Y