chore: ensure homefeed items have the same font (#4550)
### Description

The new design system fonts are ever so slightly different to the
deprecated ones. This PR prevents subtle differences in the feed items,
when the Approvals items is merged.

I also put on my cowboy hat and made some adjustments to the font
scaling behaviour for the transfer and swap feed items. i opted to
ellipsize the title and subtitle to make room for long token values on
the right hand side, also made more of an allowance for the width of the
transaction values. my goal was to stop items from becoming 3-4 lines
when either the token value was large or when the feed item description
is long. there is a detail view on tap, so i think some ellipsized
information is okay (the token values tell pretty much the whole story
anyway)

### Test plan
![Simulator Screenshot - iPhone 14 Pro - 2023-11-24 at 16 54
01](https://github.com/valora-inc/wallet/assets/20150449/3e0e422d-49fd-49d6-8462-8ef2ba346d87)

With font scaling, to show the ellipsis behaviour
![Simulator Screenshot - iPhone 14 Pro - 2023-11-24 at 16 53
37](https://github.com/valora-inc/wallet/assets/20150449/5782e424-81cb-42a5-8503-f4f5287f4787)


### Related issues

Related to RET-903

### Backwards compatibility

Y

Co-authored-by: Valora Bot <89419329+valora-bot@users.noreply.github.com>
Co-authored-by: runner <runner@Mac-1700822965867.local>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: valora-bot <valorabot@valoraapp.com>