chore: Update fiat exchange flow to use token IDs (#4450)
### Description

Changes I made:
- Updated analytics to take string rather than CiCoCurrency or
AnalyticsCurrency. Where CiCoCurrency was taken the symbol matches
exactly so no change was needed. Where AnalyticsCurrency was taken, CELO
symbol needs to be cGLD, and cREAL needs to be cReal, made
symbolToAnalyticsCurrency to do this.
- Added RawProviderQuoteWithTokenId, SimplexQuoteWithTokenId and
FiatConnectQuoteSuccessWithTokenId, which gets updated after fetching
the quotes to add in the tokenId. This way the quotes can be used
correctly on later screens with getTokenId().
- Removed all CiCoCurrency props from fiat exchange flow (excluding fiat
connect)
- Removed all Network/NetworkId props from fiat exchange flow (excluding
fiat connect)

Note: I didn't touch FiatExchangeCurrency, it is getting deleted soon
anyways and is self contained.

### Test plan

- Updated unit tests.
- Ran locally, navigated through Add flow from home screen + from asset
details page for each asset

Manual Tesing:
- Location spoofed to KE


https://github.com/valora-inc/wallet/assets/140328381/5c529c74-6246-444a-ad3f-62647c53f7fc

- Location spoofed to CA, USA


https://github.com/valora-inc/wallet/assets/140328381/da569d27-b3e6-4861-a8cf-5e5dc53e6812

### Related issues

- N/A

### Backwards compatibility

Yes

---------

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>