fix: use token ids not addresses accessing test fixture data (#4598)
### Description

A few tests imported mockTokenBalances, an object that was refactored on
9/26 to use token IDs as keys instead of addresses, but then assumed
that the keys were still addresses.

Evidently, these tests did not actually need overrides passed into
createMockStore for the tests to pass. However, I kept the overrides in
place (just fixed them) because it reduces the number of dependencies on
the exact values in our shared fixture data, which can help for creating
new tests down the road. (I don't feel too strongly about this though,
and could probably be talked into deleting the overrides instead.)

### Test plan

ci

### Related issues

na

### Backwards compatibility

yes