feat(swap): skip swap confirmation screen (#4587)
### Description

This implements the following changes in the swap flow:
- remain on the SwapScreen after tapping "Confirm Swap", showing a
loader on the button
- if the transaction(s) are successfully signed and submitted to the
network, it navigates to the home screen without waiting for them to be
confirmed. The status (pending, success, failed) can be seen from the
home screen.
- if there's an error while signing/submitting, an inline error message
is shown
- the error is dismissed if the user changes the input, or submits again

The SwapScreen and state were updated to be able to handle sending
different swaps in parallel. Each swap gets a swapId which is a `UUID`,
so the screen only displays relevant information and doesn't mix it with
other swaps possibly happening in parallel.

Note: I'll delete the `SwapExecuteScreen` in a separate PR.

### Test plan

- Updated tests
- Manually tested:


https://github.com/valora-inc/wallet/assets/57791/532fb9e2-3b43-4972-8460-5b3ab1c1715f

- Inline error (simulated by throwing while submitting):


https://github.com/valora-inc/wallet/assets/57791/c861a87a-97ca-4c58-a9b6-02d11fa4e6d2

### Related issues

- Fixes RET-910

### Backwards compatibility

Yes