feat(token-import): Validate token address (#4576)
### Description

- Once the user has entered an address for a token **and pressed**
"search" on the keyboard (or pressed "Paste" button), address validation
is performed. Note: when the address field is input, it won't
automatically start validation process.
- When the import button is pressed, the address is converted to a
tokenId. In
[RET-891/save-and-display-custom-tokens](https://linear.app/valora/issue/RET-891/save-and-display-custom-tokens)
it will be saved to the Redux state.

### Test plan

User prevented from importing (Import button is disabled), when token
address
- Already supported token, e.g.
`0x874069fa1eb16d44d622f2e0ca25eea172369bc1` (cUSD in Alfajores)
- ~Already imported token~
- Invalid EVM address

~Manually run the app to import a token address, and then try to
reimport it to see error message. This will verify that the state has
been updated.~

### Related issues

- validates and saves to the state the address for
[RET-892/lookup-contract-info](https://linear.app/valora/issue/RET-892/lookup-contract-info).
In a following PR the contract will be fetched and validated as well.

### Backwards compatibility

Yes.

---------

Co-authored-by: Kathy Luo <kathyluo18@gmail.com>