fix(viem): Fix Viem gas estimation bug (#4623)
### Description

Previously, in cases where we were attempting to estimate gas for a TX
where `gas + amount > balance`, we would _throw_ rather than returning a
null estimation. This had the downstream effect of causing the gas
estimation on the SendEnterAmount screen to spin forever, _without_ a
warning to the user telling them to enter a lower amount. This change
catches that case, and sets the gas estimate to null, causing the
appropriate warning to appear to the user.

### Test plan

Manual and unit tested. See video below:


https://github.com/valora-inc/wallet/assets/569401/e012637f-45fd-4ad6-87a8-03cf8319d0b5


### Related issues

N/A

### Backwards compatibility

Yes