feat: add automatic fee currency selection to walletconnect viem flow (#4621)
### Description

This PR:
1. always create prepared transactions on receiving an action request
from walletconnect, pass the prepared transactions to the ActionRequest
display component. always calculate the gas and feeCurrency for celo
network transactions.
2. if using viem, use the prepared transaction to show a warning if the
user does not have enough balance to pay for gas. we can also use the
prepared transaction to show the gas fees later.
3. on accepting the action request, both the walletconnect payload _and_
prepared transactions are passed to the saga that handles submitting the
transactions to the network. if using viem, it is expected that the
prepared transactions are there and can be directly sent to the network.
if not using viem, the prepared transactions are ignored.


### Test plan

I am able to make a dapp transaction on Celo and the fee currency is
paid in a non native currency.


https://github.com/valora-inc/wallet/assets/20150449/41086562-6ba3-49ad-829d-d9d7a184b634


### Related issues

- Fixes RET-905

### Backwards compatibility

Y